refactor: 컴포넌트 / 뷰 분리
